requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.3.0
torch==2.3.0
transformers==4.44.0
accelerate==0.33.0
numpy==1.26.4
huggingface_hub
sentencepiece